Optimized object type checking in C#
Refactored the way the `value` object's type is checked. Previously, the type was checked by comparing the string representation of `value.ClassInfo` to a hardcoded string. This method was replaced with a more efficient and reliable approach using the `GetType()` method and `typeof()` operator for direct type comparison. This change enhances both the reliability and performance of the type checking in the code.