Refactor report handling and update variable names
In `FinanceDocumentMaster.cs`, the usage of `LogicPOS.Reporting.Reports.Documents` has been removed and the initialization and usage of `financeMasters` has been updated. The `ticketTable` is now reset for each item in `_financeDetailsList`. In `ReportList.cs`, the class `ReportList<T>` has been updated to no longer require `T` to be a subclass of `FRBOBaseObject`. The variable `_objectHaveAttributes` has been renamed to `_objectHasAttributes` and the `List` property is now read-only. In `ReportMapping.cs`, the methods `GetFinanceDetailReportDto` and `GetFinanceMasterTotalViewReportDto` have been updated to take new parameters. Checks for `_objectHaveAttributes` have been replaced with checks for `_objectHasAttributes` in `ReportList.cs`.