Add early return to prevent UI parts from being disabled
Added a `return` statement at the beginning of the `UpdateWorkSessionUI` method in `POSMainWindow.Functions.cs` to prevent the method from executing further. This change is intended to stop certain parts of the user interface from being disabled, as indicated by the comment `//tchial0: prevent disabling some parts of the UI`.