Remove entire content of README.md
The README.md file, which contained detailed information about the LogicPOS OpenSource project, has been completely removed. This included the project title, links to manuals in Portuguese and English, a list of technologies used in the project, tools and IDEs required, a detailed directory structure, and step-by-step instructions for setting up the development environment and running the project in Visual Studio 2019.