Replace PDF viewer and remove old viewer files
Replaced the usage of `LogicPOS.PDFViewer.Winforms.PDFViewer` with `LogicPOSPDFViewer` in `DocumentPdfUtils.cs`. This change simplifies the code by removing the `using` block and directly calling `ShowDialog()` on the new viewer instance in the `ViewPdf` and `ViewReceiptPdf` methods.

Removed `PdfViewer.Designer.cs` and `PdfViewer.cs` files, which contained the partial class definition and implementation of the old `PdfViewer` Windows Form.

Introduced a new PDF viewer form `LogicPOSPDFViewer` using the `PdfiumViewer` library. The new form is designed to display PDF documents and includes a constructor to initialize the form and load the PDF document when the form is loaded.