Add report of total sold by item
Implemented classes and methods to generate item sales reports, including `GetArticleTotalSoldReportPdfQuery` and its handler. Added a button in the UI to access the report and implemented the `BtnArticleTotalSoldReport_Clicked` event to open a date filter modal. Updated the `ReportsService` service with the `ShowArticleTotalSoldReport` method. Modified the `LogicPOS.UI.csproj` project file to include new code files.