Added date validation and formatting improvements
Validation features were implemented in the `TxtStartDate` and `TxtEndDate` fields across multiple files, ensuring valid dates are entered and automatically adjusting to the last day of the month when necessary.

The regular expression for date validation was improved to only accept the `yyyy-MM-dd` format for years between 1800 and 2099.

Quantity formatting in `TreeView` columns was changed to display two decimal places, improving data presentation.

The `DateTimeExtension` class was created with the `ValidateDate` extension method to facilitate date validation and formatting.

Imports of `LogicPOS.UI.Extensions` were added to several files to utilize the new features.

The `StartDate` property in the `ReportsFilterModal` class was changed to `get;`. set;`, allowing more flexible manipulation of the start date value.