Disables filter buttons and improves modals and structure
Added the `DisableFilterButton` functionality to several pages, disabling filter buttons where they are not needed. Implemented the method as virtual in the base class `Page` to allow overriding.

Updates to the `AgtSeriesFilterModal` modal include validation adjustments, removal of selection buttons, and addition of the `Clear` method to reset fields. Changed the modal title to ‘Filter Online Series - AGT’.

Included new files and components in the project (`LogicPOS.UI.csproj`), such as licensing services, utilities, and modals. Removed redundant methods and unnecessary references, as well as general improvements to code organisation and structure.