Added series communication button to the modal
The `_btnATSeriesCommunicate` button has been added to the `DocumentSerieModal` modal, including its creation, initialization, and display in the layout. The `CreateIconButton` method has been implemented to standardize the creation of buttons with icons. The sensitivity of the `_txtATDocCodeValidationSerie` component has been adjusted to always be active. The logic in `DocumentSeriesPage` has been changed to condition the display mode if the default country code is not "PT".